chore: restructure code
