+++ pkg/commands/drift/assets/assets.go
