+++ pkg/commands/drift/assets/assets_test.go
