+++ pkg/commands/drift/drift.go
@@ -21,9 +21,9 @@ import (
 
 	"github.com/abcxyz/pkg/logging"
 
-	"github.com/abcxyz/guardian/pkg/drift/assets"
-	"github.com/abcxyz/guardian/pkg/drift/iam"
-	"github.com/abcxyz/guardian/pkg/drift/terraform"
+	"github.com/abcxyz/guardian/pkg/commands/drift/assets"
+	"github.com/abcxyz/guardian/pkg/commands/drift/iam"
+	"github.com/abcxyz/guardian/pkg/commands/drift/terraform"
 )
 
 // IAMDrift represents the detected iam drift in a gcp org.
