+++ pkg/commands/drift/terraform/terraform.go
@@ -22,9 +22,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/abcxyz/guardian/pkg/drift/assets"
-	"github.com/abcxyz/guardian/pkg/drift/gcs"
-	"github.com/abcxyz/guardian/pkg/drift/iam"
+	"github.com/abcxyz/guardian/pkg/commands/drift/assets"
+	"github.com/abcxyz/guardian/pkg/commands/drift/gcs"
+	"github.com/abcxyz/guardian/pkg/commands/drift/iam"
 )
 
 const (
